Add DER to PEM test
